Begin implementing elba tool and add benchmarks
This is the start of the port of ELBA from the hornnonlin implementation 
to horn. Currently it adds the precond to the first CHC, constructs 
paths, and uses the datalearner. There is still progress to be made on 
handling sequencial loops. There is an alert which identifies 
multiple/single loops, but this is crude and more progress is needed.